\name{rworldxtra-package}
\alias{rworldxtra-package}
\alias{rworldxtra}
\docType{package}
\title{
For mapping global data.
}
\description{
Enables mapping of country level and gridded user datasets by facilitating joining to world maps and visualisation options.  
}
\details{
\tabular{ll}{
Package: \tab rworldxtra\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2012-10-1\cr
License: \tab GPL (>= 2)\cr
}

Version 1.01 newly uses updated Natural Earth Data for country boundaries.


}
\author{
Andy South

Maintainer: <southandy@gmail.com>
}
\references{
Derived from : http://www.naturalearthdata.com/downloads/10m-cultural-vectors/

}

\keyword{ package }


\examples{

data(countriesHigh)            


}
